requirements.txt
ultralytics>=8.1.0,<9.0.0
opencv-python>=4.5.0,<5.0.0
numpy>=1.21.0,<2.0.0
pillow>=8.3.0,<10.0.0
matplotlib>=3.3.0,<4.0.0
fastapi>=0.68.0,<1.0.0
uvicorn>=0.15.0,<1.0.0
python-multipart>=0.0.5,<1.0.0